move some functionality from ChainBase to Chain
